Assets and basic Text
Added a few new assets such as the main menu background and the font for text. Made a new window handler so that the images of the ship and the actual descriptions and action handling can be done in seperate windows without cluttering one window. Made a text handlet with some sample text for each 'Scene' of the spaceship. Next up is keyboard handling and input.